# shared/api_contract.json
{"openapi":"3.1.0","info":{"title":"Olist Marketplace Platform","description":"Olist Marketplace end-to-end analytics and review-sentiment intelligence API.","version":"1.0.0"},"paths":{"/api/v1/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/models/status":{"get":{"tags":["models"],"summary":"Model Status","operationId":"model_status_api_v1_models_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/models/info":{"get":{"tags":["models"],"summary":"Model Info","operationId":"model_info_api_v1_models_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/sentiment/predict":{"post":{"tags":["sentiment"],"summary":"Predict","operationId":"predict_api_v1_sentiment_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SentimentPredictionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sentiment/predict-batch":{"post":{"tags":["sentiment"],"summary":"Predict Batch","operationId":"predict_batch_api_v1_sentiment_predict_batch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchPredictionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/summary":{"get":{"tags":["analytics"],"summary":"Summary","operationId":"summary_api_v1_analytics_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analytics/orders/monthly":{"get":{"tags":["analytics"],"summary":"Monthly Orders","operationId":"monthly_orders_api_v1_analytics_orders_monthly_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analytics/revenue/monthly":{"get":{"tags":["analytics"],"summary":"Monthly Revenue","operationId":"monthly_revenue_api_v1_analytics_revenue_monthly_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analytics/reviews/distribution":{"get":{"tags":["analytics"],"summary":"Review Distribution","operationId":"review_distribution_api_v1_analytics_reviews_distribution_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analytics/delivery/summary":{"get":{"tags":["analytics"],"summary":"Delivery Summary","operationId":"delivery_summary_api_v1_analytics_delivery_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analytics/payments/distribution":{"get":{"tags":["analytics"],"summary":"Payment Distribution","operationId":"payment_distribution_api_v1_analytics_payments_distribution_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/customers/summary":{"get":{"tags":["customers"],"summary":"Customer Summary","operationId":"customer_summary_api_v1_customers_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/customers/top-cities":{"get":{"tags":["customers"],"summary":"Top Cities","operationId":"top_cities_api_v1_customers_top_cities_get","parameters":[{"name":"n","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"N"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/customers/segments":{"get":{"tags":["customers"],"summary":"Segments","operationId":"segments_api_v1_customers_segments_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/customers/segments/{segment_name}":{"get":{"tags":["customers"],"summary":"Segment Detail","operationId":"segment_detail_api_v1_customers_segments__segment_name__get","parameters":[{"name":"segment_name","in":"path","required":true,"schema":{"type":"string","title":"Segment Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sellers/summary":{"get":{"tags":["sellers"],"summary":"Seller Summary","operationId":"seller_summary_api_v1_sellers_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/sellers/performance":{"get":{"tags":["sellers"],"summary":"Seller Performance","operationId":"seller_performance_api_v1_sellers_performance_get","parameters":[{"name":"n","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"N"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/products/categories":{"get":{"tags":["products"],"summary":"Categories","operationId":"categories_api_v1_products_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/products/category-performance":{"get":{"tags":["products"],"summary":"Category Performance","operationId":"category_performance_api_v1_products_category_performance_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/geography/state-performance":{"get":{"tags":["geography"],"summary":"State Performance","operationId":"state_performance_api_v1_geography_state_performance_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/segmentation/rfm-summary":{"get":{"tags":["segmentation"],"summary":"Rfm Summary","operationId":"rfm_summary_api_v1_segmentation_rfm_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/segmentation/predict":{"post":{"tags":["segmentation"],"summary":"Predict","operationId":"predict_api_v1_segmentation_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RfmPredictionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BatchPredictionItem":{"properties":{"id":{"type":"string","title":"Id"},"text":{"type":"string","maxLength":2000,"minLength":1,"title":"Text"}},"type":"object","required":["id","text"],"title":"BatchPredictionItem"},"BatchPredictionRequest":{"properties":{"items":{"items":{"$ref":"#/components/schemas/BatchPredictionItem"},"type":"array","maxItems":64,"minItems":1,"title":"Items"},"model_name":{"type":"string","enum":["bert","cnn2d"],"title":"Model Name","default":"bert"}},"type":"object","required":["items"],"title":"BatchPredictionRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RfmPredictionRequest":{"properties":{"recency":{"type":"number","minimum":0.0,"title":"Recency","description":"Days since the customer's last order"},"frequency":{"type":"number","minimum":1.0,"title":"Frequency","description":"Number of distinct orders"},"monetary":{"type":"number","minimum":0.0,"title":"Monetary","description":"Total spend across orders (BRL)"}},"type":"object","required":["recency","frequency","monetary"],"title":"RfmPredictionRequest"},"SentimentPredictionRequest":{"properties":{"text":{"type":"string","maxLength":2000,"minLength":1,"title":"Text"},"model_name":{"type":"string","enum":["bert","cnn2d"],"title":"Model Name","default":"bert"},"source_language":{"type":"string","enum":["en","pt"],"title":"Source Language","default":"en"},"translate":{"type":"boolean","title":"Translate","default":false}},"type":"object","required":["text"],"title":"SentimentPredictionRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}